policy

+++ app/Http/Controllers/JabatanController.php
@@ -20,6 +20,7 @@ class JabatanController extends Controller
     public function index()
     {
         // $this->authorize('create',Jabatan::class);
+        // $this->authorize('create',Jabatan::class);
         $nomor = 1;
         $jabatan = Jabatan::all();
         return view('jabatan.index',compact('nomor','jabatan'));
@@ -43,7 +44,7 @@ class JabatanController extends Controller
      */
     public function store(Request $request)
     {
-        
+        $this->authorize('create',Jabatan::class);
         $jab = new Jabatan;
 
         $jab->kode = $request->kode;
@@ -103,6 +104,7 @@ class JabatanController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('create',$jabatan);
         $jab = Jabatan::find($id);
         $jab->delete();
